factor out dt
