+++ narwhals/expr.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 from typing import Any
 from typing import Callable
-from typing import Generic
 from typing import Iterable
 from typing import Literal
 from typing import Mapping
@@ -12,6 +11,11 @@ from typing import TypeVar
 
 from narwhals.dependencies import is_numpy_array
 from narwhals.dtypes import _validate_dtype
+from narwhals.expr_cat import ExprCatNamespace
+from narwhals.expr_dt import ExprDateTimeNamespace
+from narwhals.expr_list import ExprListNamespace
+from narwhals.expr_name import ExprNameNamespace
+from narwhals.expr_str import ExprStringNamespace
 from narwhals.utils import _validate_rolling_arguments
 from narwhals.utils import flatten
 
@@ -23,6 +27,8 @@ if TYPE_CHECKING:
     from narwhals.typing import CompliantNamespace
     from narwhals.typing import CompliantSeriesT_co
     from narwhals.typing import IntoExpr
+
+ExprT = TypeVar("ExprT", bound="Expr")
 
 
 def extract_compliant(
@@ -4230,2694 +4236,6 @@ class Expr:
     @property
     def list(self: Self) -> ExprListNamespace[Self]:
         return ExprListNamespace(self)
-
-
-ExprT = TypeVar("ExprT", bound=Expr)
-
-
-class ExprCatNamespace(Generic[ExprT]):
-    def __init__(self: Self, expr: ExprT) -> None:
-        self._expr = expr
-
-    def get_categories(self: Self) -> ExprT:
-        """Get unique categories from column.
-
-        Returns:
-            A new expression.
-
-        Examples:
-            Let's create some dataframes:
-
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {"fruits": ["apple", "mango", "mango"]}
-            >>> df_pd = pd.DataFrame(data, dtype="category")
-            >>> df_pl = pl.DataFrame(data, schema={"fruits": pl.Categorical})
-
-            We define a dataframe-agnostic function to get unique categories
-            from column 'fruits':
-
-            >>> def agnostic_cat_get_categories(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.select(nw.col("fruits").cat.get_categories()).to_native()
-
-           We can then pass any supported library such as pandas or Polars to
-           `agnostic_cat_get_categories`:
-
-            >>> agnostic_cat_get_categories(df_pd)
-              fruits
-            0  apple
-            1  mango
-
-            >>> agnostic_cat_get_categories(df_pl)
-            shape: (2, 1)
-            ┌────────┐
-            │ fruits │
-            │ ---    │
-            │ str    │
-            ╞════════╡
-            │ apple  │
-            │ mango  │
-            └────────┘
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).cat.get_categories()
-        )
-
-
-class ExprStringNamespace(Generic[ExprT]):
-    def __init__(self: Self, expr: ExprT) -> None:
-        self._expr = expr
-
-    def len_chars(self: Self) -> ExprT:
-        r"""Return the length of each string as the number of characters.
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {"words": ["foo", "Café", "345", "東京", None]}
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_str_len_chars(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         words_len=nw.col("words").str.len_chars()
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_str_len_chars`:
-
-            >>> agnostic_str_len_chars(df_pd)
-              words  words_len
-            0   foo        3.0
-            1  Café        4.0
-            2   345        3.0
-            3    東京        2.0
-            4  None        NaN
-
-            >>> agnostic_str_len_chars(df_pl)
-            shape: (5, 2)
-            ┌───────┬───────────┐
-            │ words ┆ words_len │
-            │ ---   ┆ ---       │
-            │ str   ┆ u32       │
-            ╞═══════╪═══════════╡
-            │ foo   ┆ 3         │
-            │ Café  ┆ 4         │
-            │ 345   ┆ 3         │
-            │ 東京  ┆ 2         │
-            │ null  ┆ null      │
-            └───────┴───────────┘
-
-            >>> agnostic_str_len_chars(df_pa)
-            pyarrow.Table
-            words: string
-            words_len: int32
-            ----
-            words: [["foo","Café","345","東京",null]]
-            words_len: [[3,4,3,2,null]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).str.len_chars()
-        )
-
-    def replace(
-        self, pattern: str, value: str, *, literal: bool = False, n: int = 1
-    ) -> ExprT:
-        r"""Replace first matching regex/literal substring with a new string value.
-
-        Arguments:
-            pattern: A valid regular expression pattern.
-            value: String that will replace the matched substring.
-            literal: Treat `pattern` as a literal string.
-            n: Number of matches to replace.
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {"foo": ["123abc", "abc abc123"]}
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_str_replace(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     df = df.with_columns(replaced=nw.col("foo").str.replace("abc", ""))
-            ...     return df.to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_str_replace`:
-
-            >>> agnostic_str_replace(df_pd)
-                      foo replaced
-            0      123abc      123
-            1  abc abc123   abc123
-
-            >>> agnostic_str_replace(df_pl)
-            shape: (2, 2)
-            ┌────────────┬──────────┐
-            │ foo        ┆ replaced │
-            │ ---        ┆ ---      │
-            │ str        ┆ str      │
-            ╞════════════╪══════════╡
-            │ 123abc     ┆ 123      │
-            │ abc abc123 ┆  abc123  │
-            └────────────┴──────────┘
-
-            >>> agnostic_str_replace(df_pa)
-            pyarrow.Table
-            foo: string
-            replaced: string
-            ----
-            foo: [["123abc","abc abc123"]]
-            replaced: [["123"," abc123"]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).str.replace(
-                pattern, value, literal=literal, n=n
-            )
-        )
-
-    def replace_all(
-        self: Self, pattern: str, value: str, *, literal: bool = False
-    ) -> ExprT:
-        r"""Replace all matching regex/literal substring with a new string value.
-
-        Arguments:
-            pattern: A valid regular expression pattern.
-            value: String that will replace the matched substring.
-            literal: Treat `pattern` as a literal string.
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {"foo": ["123abc", "abc abc123"]}
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_str_replace_all(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     df = df.with_columns(replaced=nw.col("foo").str.replace_all("abc", ""))
-            ...     return df.to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_str_replace_all`:
-
-            >>> agnostic_str_replace_all(df_pd)
-                      foo replaced
-            0      123abc      123
-            1  abc abc123      123
-
-            >>> agnostic_str_replace_all(df_pl)
-            shape: (2, 2)
-            ┌────────────┬──────────┐
-            │ foo        ┆ replaced │
-            │ ---        ┆ ---      │
-            │ str        ┆ str      │
-            ╞════════════╪══════════╡
-            │ 123abc     ┆ 123      │
-            │ abc abc123 ┆  123     │
-            └────────────┴──────────┘
-
-            >>> agnostic_str_replace_all(df_pa)
-            pyarrow.Table
-            foo: string
-            replaced: string
-            ----
-            foo: [["123abc","abc abc123"]]
-            replaced: [["123"," 123"]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).str.replace_all(
-                pattern, value, literal=literal
-            )
-        )
-
-    def strip_chars(self: Self, characters: str | None = None) -> ExprT:
-        r"""Remove leading and trailing characters.
-
-        Arguments:
-            characters: The set of characters to be removed. All combinations of this
-                set of characters will be stripped from the start and end of the string.
-                If set to None (default), all leading and trailing whitespace is removed
-                instead.
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> from typing import Any
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrame
-            >>>
-            >>> data = {"fruits": ["apple", "\nmango"]}
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_str_strip_chars(df_native: IntoFrame) -> dict[str, Any]:
-            ...     df = nw.from_native(df_native)
-            ...     df = df.with_columns(stripped=nw.col("fruits").str.strip_chars())
-            ...     return df.to_dict(as_series=False)
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_str_strip_chars`:
-
-            >>> agnostic_str_strip_chars(df_pd)
-            {'fruits': ['apple', '\nmango'], 'stripped': ['apple', 'mango']}
-
-            >>> agnostic_str_strip_chars(df_pl)
-            {'fruits': ['apple', '\nmango'], 'stripped': ['apple', 'mango']}
-
-            >>> agnostic_str_strip_chars(df_pa)
-            {'fruits': ['apple', '\nmango'], 'stripped': ['apple', 'mango']}
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).str.strip_chars(characters)
-        )
-
-    def starts_with(self: Self, prefix: str) -> ExprT:
-        r"""Check if string values start with a substring.
-
-        Arguments:
-            prefix: prefix substring
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {"fruits": ["apple", "mango", None]}
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_str_starts_with(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         has_prefix=nw.col("fruits").str.starts_with("app")
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_str_starts_with`:
-
-            >>> agnostic_str_starts_with(df_pd)
-              fruits has_prefix
-            0  apple       True
-            1  mango      False
-            2   None       None
-
-            >>> agnostic_str_starts_with(df_pl)
-            shape: (3, 2)
-            ┌────────┬────────────┐
-            │ fruits ┆ has_prefix │
-            │ ---    ┆ ---        │
-            │ str    ┆ bool       │
-            ╞════════╪════════════╡
-            │ apple  ┆ true       │
-            │ mango  ┆ false      │
-            │ null   ┆ null       │
-            └────────┴────────────┘
-
-            >>> agnostic_str_starts_with(df_pa)
-            pyarrow.Table
-            fruits: string
-            has_prefix: bool
-            ----
-            fruits: [["apple","mango",null]]
-            has_prefix: [[true,false,null]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).str.starts_with(prefix)
-        )
-
-    def ends_with(self: Self, suffix: str) -> ExprT:
-        r"""Check if string values end with a substring.
-
-        Arguments:
-            suffix: suffix substring
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {"fruits": ["apple", "mango", None]}
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_str_ends_with(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         has_suffix=nw.col("fruits").str.ends_with("ngo")
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_str_ends_with`:
-
-            >>> agnostic_str_ends_with(df_pd)
-              fruits has_suffix
-            0  apple      False
-            1  mango       True
-            2   None       None
-
-            >>> agnostic_str_ends_with(df_pl)
-            shape: (3, 2)
-            ┌────────┬────────────┐
-            │ fruits ┆ has_suffix │
-            │ ---    ┆ ---        │
-            │ str    ┆ bool       │
-            ╞════════╪════════════╡
-            │ apple  ┆ false      │
-            │ mango  ┆ true       │
-            │ null   ┆ null       │
-            └────────┴────────────┘
-
-            >>> agnostic_str_ends_with(df_pa)
-            pyarrow.Table
-            fruits: string
-            has_suffix: bool
-            ----
-            fruits: [["apple","mango",null]]
-            has_suffix: [[false,true,null]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).str.ends_with(suffix)
-        )
-
-    def contains(self: Self, pattern: str, *, literal: bool = False) -> ExprT:
-        r"""Check if string contains a substring that matches a pattern.
-
-        Arguments:
-            pattern: A Character sequence or valid regular expression pattern.
-            literal: If True, treats the pattern as a literal string.
-                     If False, assumes the pattern is a regular expression.
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {"pets": ["cat", "dog", "rabbit and parrot", "dove", None]}
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_str_contains(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         default_match=nw.col("pets").str.contains("parrot|Dove"),
-            ...         case_insensitive_match=nw.col("pets").str.contains("(?i)parrot|Dove"),
-            ...         literal_match=nw.col("pets").str.contains(
-            ...             "parrot|Dove", literal=True
-            ...         ),
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_str_contains`:
-
-            >>> agnostic_str_contains(df_pd)
-                            pets default_match case_insensitive_match literal_match
-            0                cat         False                  False         False
-            1                dog         False                  False         False
-            2  rabbit and parrot          True                   True         False
-            3               dove         False                   True         False
-            4               None          None                   None          None
-
-            >>> agnostic_str_contains(df_pl)
-            shape: (5, 4)
-            ┌───────────────────┬───────────────┬────────────────────────┬───────────────┐
-            │ pets              ┆ default_match ┆ case_insensitive_match ┆ literal_match │
-            │ ---               ┆ ---           ┆ ---                    ┆ ---           │
-            │ str               ┆ bool          ┆ bool                   ┆ bool          │
-            ╞═══════════════════╪═══════════════╪════════════════════════╪═══════════════╡
-            │ cat               ┆ false         ┆ false                  ┆ false         │
-            │ dog               ┆ false         ┆ false                  ┆ false         │
-            │ rabbit and parrot ┆ true          ┆ true                   ┆ false         │
-            │ dove              ┆ false         ┆ true                   ┆ false         │
-            │ null              ┆ null          ┆ null                   ┆ null          │
-            └───────────────────┴───────────────┴────────────────────────┴───────────────┘
-
-            >>> agnostic_str_contains(df_pa)
-            pyarrow.Table
-            pets: string
-            default_match: bool
-            case_insensitive_match: bool
-            literal_match: bool
-            ----
-            pets: [["cat","dog","rabbit and parrot","dove",null]]
-            default_match: [[false,false,true,false,null]]
-            case_insensitive_match: [[false,false,true,true,null]]
-            literal_match: [[false,false,false,false,null]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).str.contains(
-                pattern, literal=literal
-            )
-        )
-
-    def slice(self: Self, offset: int, length: int | None = None) -> ExprT:
-        r"""Create subslices of the string values of an expression.
-
-        Arguments:
-            offset: Start index. Negative indexing is supported.
-            length: Length of the slice. If set to `None` (default), the slice is taken to the
-                end of the string.
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {"s": ["pear", None, "papaya", "dragonfruit"]}
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_str_slice(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         s_sliced=nw.col("s").str.slice(4, length=3)
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_str_slice`:
-
-            >>> agnostic_str_slice(df_pd)  # doctest: +NORMALIZE_WHITESPACE
-                         s s_sliced
-            0         pear
-            1         None     None
-            2       papaya       ya
-            3  dragonfruit      onf
-
-            >>> agnostic_str_slice(df_pl)
-            shape: (4, 2)
-            ┌─────────────┬──────────┐
-            │ s           ┆ s_sliced │
-            │ ---         ┆ ---      │
-            │ str         ┆ str      │
-            ╞═════════════╪══════════╡
-            │ pear        ┆          │
-            │ null        ┆ null     │
-            │ papaya      ┆ ya       │
-            │ dragonfruit ┆ onf      │
-            └─────────────┴──────────┘
-
-            >>> agnostic_str_slice(df_pa)
-            pyarrow.Table
-            s: string
-            s_sliced: string
-            ----
-            s: [["pear",null,"papaya","dragonfruit"]]
-            s_sliced: [["",null,"ya","onf"]]
-
-            Using negative indexes:
-
-            >>> def agnostic_str_slice_negative(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(s_sliced=nw.col("s").str.slice(-3)).to_native()
-
-            >>> agnostic_str_slice_negative(df_pd)
-                         s s_sliced
-            0         pear      ear
-            1         None     None
-            2       papaya      aya
-            3  dragonfruit      uit
-
-            >>> agnostic_str_slice_negative(df_pl)
-            shape: (4, 2)
-            ┌─────────────┬──────────┐
-            │ s           ┆ s_sliced │
-            │ ---         ┆ ---      │
-            │ str         ┆ str      │
-            ╞═════════════╪══════════╡
-            │ pear        ┆ ear      │
-            │ null        ┆ null     │
-            │ papaya      ┆ aya      │
-            │ dragonfruit ┆ uit      │
-            └─────────────┴──────────┘
-
-            >>> agnostic_str_slice_negative(df_pa)
-            pyarrow.Table
-            s: string
-            s_sliced: string
-            ----
-            s: [["pear",null,"papaya","dragonfruit"]]
-            s_sliced: [["ear",null,"aya","uit"]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).str.slice(
-                offset=offset, length=length
-            )
-        )
-
-    def head(self: Self, n: int = 5) -> ExprT:
-        r"""Take the first n elements of each string.
-
-        Arguments:
-            n: Number of elements to take. Negative indexing is **not** supported.
-
-        Returns:
-            A new expression.
-
-        Notes:
-            If the length of the string has fewer than `n` characters, the full string is returned.
-
-        Examples:
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {"lyrics": ["Atatata", "taata", "taatatata", "zukkyun"]}
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_str_head(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         lyrics_head=nw.col("lyrics").str.head()
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_str_head`:
-
-            >>> agnostic_str_head(df_pd)
-                  lyrics lyrics_head
-            0    Atatata       Atata
-            1      taata       taata
-            2  taatatata       taata
-            3    zukkyun       zukky
-
-            >>> agnostic_str_head(df_pl)
-            shape: (4, 2)
-            ┌───────────┬─────────────┐
-            │ lyrics    ┆ lyrics_head │
-            │ ---       ┆ ---         │
-            │ str       ┆ str         │
-            ╞═══════════╪═════════════╡
-            │ Atatata   ┆ Atata       │
-            │ taata     ┆ taata       │
-            │ taatatata ┆ taata       │
-            │ zukkyun   ┆ zukky       │
-            └───────────┴─────────────┘
-
-            >>> agnostic_str_head(df_pa)
-            pyarrow.Table
-            lyrics: string
-            lyrics_head: string
-            ----
-            lyrics: [["Atatata","taata","taatatata","zukkyun"]]
-            lyrics_head: [["Atata","taata","taata","zukky"]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).str.slice(0, n)
-        )
-
-    def tail(self: Self, n: int = 5) -> ExprT:
-        r"""Take the last n elements of each string.
-
-        Arguments:
-            n: Number of elements to take. Negative indexing is **not** supported.
-
-        Returns:
-            A new expression.
-
-        Notes:
-            If the length of the string has fewer than `n` characters, the full string is returned.
-
-        Examples:
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {"lyrics": ["Atatata", "taata", "taatatata", "zukkyun"]}
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_str_tail(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         lyrics_tail=nw.col("lyrics").str.tail()
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_str_tail`:
-
-            >>> agnostic_str_tail(df_pd)
-                  lyrics lyrics_tail
-            0    Atatata       atata
-            1      taata       taata
-            2  taatatata       atata
-            3    zukkyun       kkyun
-
-            >>> agnostic_str_tail(df_pl)
-            shape: (4, 2)
-            ┌───────────┬─────────────┐
-            │ lyrics    ┆ lyrics_tail │
-            │ ---       ┆ ---         │
-            │ str       ┆ str         │
-            ╞═══════════╪═════════════╡
-            │ Atatata   ┆ atata       │
-            │ taata     ┆ taata       │
-            │ taatatata ┆ atata       │
-            │ zukkyun   ┆ kkyun       │
-            └───────────┴─────────────┘
-
-            >>> agnostic_str_tail(df_pa)
-            pyarrow.Table
-            lyrics: string
-            lyrics_tail: string
-            ----
-            lyrics: [["Atatata","taata","taatatata","zukkyun"]]
-            lyrics_tail: [["atata","taata","atata","kkyun"]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).str.slice(
-                offset=-n, length=None
-            )
-        )
-
-    def to_datetime(self: Self, format: str | None = None) -> ExprT:  # noqa: A002
-        """Convert to Datetime dtype.
-
-        Warning:
-            As different backends auto-infer format in different ways, if `format=None`
-            there is no guarantee that the result will be equal.
-
-        Arguments:
-            format: Format to use for conversion. If set to None (default), the format is
-                inferred from the data.
-
-        Returns:
-            A new expression.
-
-        Notes:
-            pandas defaults to nanosecond time unit, Polars to microsecond.
-            Prior to pandas 2.0, nanoseconds were the only time unit supported
-            in pandas, with no ability to set any other one. The ability to
-            set the time unit in pandas, if the version permits, will arrive.
-
-        Examples:
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = ["2020-01-01", "2020-01-02"]
-            >>> df_pd = pd.DataFrame({"a": data})
-            >>> df_pl = pl.DataFrame({"a": data})
-            >>> df_pa = pa.table({"a": data})
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_str_to_datetime(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.select(
-            ...         nw.col("a").str.to_datetime(format="%Y-%m-%d")
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_str_to_datetime`:
-
-            >>> agnostic_str_to_datetime(df_pd)
-                       a
-            0 2020-01-01
-            1 2020-01-02
-
-            >>> agnostic_str_to_datetime(df_pl)
-            shape: (2, 1)
-            ┌─────────────────────┐
-            │ a                   │
-            │ ---                 │
-            │ datetime[μs]        │
-            ╞═════════════════════╡
-            │ 2020-01-01 00:00:00 │
-            │ 2020-01-02 00:00:00 │
-            └─────────────────────┘
-
-            >>> agnostic_str_to_datetime(df_pa)
-            pyarrow.Table
-            a: timestamp[us]
-            ----
-            a: [[2020-01-01 00:00:00.000000,2020-01-02 00:00:00.000000]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).str.to_datetime(format=format)
-        )
-
-    def to_uppercase(self: Self) -> ExprT:
-        r"""Transform string to uppercase variant.
-
-        Returns:
-            A new expression.
-
-        Notes:
-            The PyArrow backend will convert 'ß' to 'ẞ' instead of 'SS'.
-            For more info see [the related issue](https://github.com/apache/arrow/issues/34599).
-            There may be other unicode-edge-case-related variations across implementations.
-
-        Examples:
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {"fruits": ["apple", "mango", None]}
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_str_to_uppercase(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         upper_col=nw.col("fruits").str.to_uppercase()
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_str_to_uppercase`:
-
-            >>> agnostic_str_to_uppercase(df_pd)
-              fruits upper_col
-            0  apple     APPLE
-            1  mango     MANGO
-            2   None      None
-
-            >>> agnostic_str_to_uppercase(df_pl)
-            shape: (3, 2)
-            ┌────────┬───────────┐
-            │ fruits ┆ upper_col │
-            │ ---    ┆ ---       │
-            │ str    ┆ str       │
-            ╞════════╪═══════════╡
-            │ apple  ┆ APPLE     │
-            │ mango  ┆ MANGO     │
-            │ null   ┆ null      │
-            └────────┴───────────┘
-
-            >>> agnostic_str_to_uppercase(df_pa)
-            pyarrow.Table
-            fruits: string
-            upper_col: string
-            ----
-            fruits: [["apple","mango",null]]
-            upper_col: [["APPLE","MANGO",null]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).str.to_uppercase()
-        )
-
-    def to_lowercase(self: Self) -> ExprT:
-        r"""Transform string to lowercase variant.
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {"fruits": ["APPLE", "MANGO", None]}
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_str_to_lowercase(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         lower_col=nw.col("fruits").str.to_lowercase()
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_str_to_lowercase`:
-
-            >>> agnostic_str_to_lowercase(df_pd)
-              fruits lower_col
-            0  APPLE     apple
-            1  MANGO     mango
-            2   None      None
-
-            >>> agnostic_str_to_lowercase(df_pl)
-            shape: (3, 2)
-            ┌────────┬───────────┐
-            │ fruits ┆ lower_col │
-            │ ---    ┆ ---       │
-            │ str    ┆ str       │
-            ╞════════╪═══════════╡
-            │ APPLE  ┆ apple     │
-            │ MANGO  ┆ mango     │
-            │ null   ┆ null      │
-            └────────┴───────────┘
-
-            >>> agnostic_str_to_lowercase(df_pa)
-            pyarrow.Table
-            fruits: string
-            lower_col: string
-            ----
-            fruits: [["APPLE","MANGO",null]]
-            lower_col: [["apple","mango",null]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).str.to_lowercase()
-        )
-
-
-class ExprDateTimeNamespace(Generic[ExprT]):
-    def __init__(self: Self, expr: ExprT) -> None:
-        self._expr = expr
-
-    def date(self: Self) -> ExprT:
-        """Extract the date from underlying DateTime representation.
-
-        Returns:
-            A new expression.
-
-        Raises:
-            NotImplementedError: If pandas default backend is being used.
-
-        Examples:
-            >>> from datetime import datetime
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {"a": [datetime(2012, 1, 7, 10, 20), datetime(2023, 3, 10, 11, 32)]}
-            >>> df_pd = pd.DataFrame(data).convert_dtypes(dtype_backend="pyarrow")
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a library agnostic function:
-
-            >>> def agnostic_dt_date(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.select(nw.col("a").dt.date()).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_dt_date`:
-
-            >>> agnostic_dt_date(df_pd)
-                        a
-            0  2012-01-07
-            1  2023-03-10
-
-            >>> agnostic_dt_date(df_pl)
-            shape: (2, 1)
-            ┌────────────┐
-            │ a          │
-            │ ---        │
-            │ date       │
-            ╞════════════╡
-            │ 2012-01-07 │
-            │ 2023-03-10 │
-            └────────────┘
-
-            >>> agnostic_dt_date(df_pa)
-            pyarrow.Table
-            a: date32[day]
-            ----
-            a: [[2012-01-07,2023-03-10]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).dt.date()
-        )
-
-    def year(self: Self) -> ExprT:
-        """Extract year from underlying DateTime representation.
-
-        Returns the year number in the calendar date.
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> from datetime import datetime
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {
-            ...     "datetime": [
-            ...         datetime(1978, 6, 1),
-            ...         datetime(2024, 12, 13),
-            ...         datetime(2065, 1, 1),
-            ...     ]
-            ... }
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_dt_year(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         nw.col("datetime").dt.year().alias("year")
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_dt_year`:
-
-            >>> agnostic_dt_year(df_pd)
-                datetime  year
-            0 1978-06-01  1978
-            1 2024-12-13  2024
-            2 2065-01-01  2065
-
-            >>> agnostic_dt_year(df_pl)
-            shape: (3, 2)
-            ┌─────────────────────┬──────┐
-            │ datetime            ┆ year │
-            │ ---                 ┆ ---  │
-            │ datetime[μs]        ┆ i32  │
-            ╞═════════════════════╪══════╡
-            │ 1978-06-01 00:00:00 ┆ 1978 │
-            │ 2024-12-13 00:00:00 ┆ 2024 │
-            │ 2065-01-01 00:00:00 ┆ 2065 │
-            └─────────────────────┴──────┘
-
-            >>> agnostic_dt_year(df_pa)
-            pyarrow.Table
-            datetime: timestamp[us]
-            year: int64
-            ----
-            datetime: [[1978-06-01 00:00:00.000000,2024-12-13 00:00:00.000000,2065-01-01 00:00:00.000000]]
-            year: [[1978,2024,2065]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).dt.year()
-        )
-
-    def month(self: Self) -> ExprT:
-        """Extract month from underlying DateTime representation.
-
-        Returns the month number starting from 1. The return value ranges from 1 to 12.
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> from datetime import datetime
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {
-            ...     "datetime": [
-            ...         datetime(1978, 6, 1),
-            ...         datetime(2024, 12, 13),
-            ...         datetime(2065, 1, 1),
-            ...     ]
-            ... }
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_dt_month(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         nw.col("datetime").dt.month().alias("month"),
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_dt_month`:
-
-            >>> agnostic_dt_month(df_pd)
-                datetime  month
-            0 1978-06-01      6
-            1 2024-12-13     12
-            2 2065-01-01      1
-
-            >>> agnostic_dt_month(df_pl)
-            shape: (3, 2)
-            ┌─────────────────────┬───────┐
-            │ datetime            ┆ month │
-            │ ---                 ┆ ---   │
-            │ datetime[μs]        ┆ i8    │
-            ╞═════════════════════╪═══════╡
-            │ 1978-06-01 00:00:00 ┆ 6     │
-            │ 2024-12-13 00:00:00 ┆ 12    │
-            │ 2065-01-01 00:00:00 ┆ 1     │
-            └─────────────────────┴───────┘
-
-            >>> agnostic_dt_month(df_pa)
-            pyarrow.Table
-            datetime: timestamp[us]
-            month: int64
-            ----
-            datetime: [[1978-06-01 00:00:00.000000,2024-12-13 00:00:00.000000,2065-01-01 00:00:00.000000]]
-            month: [[6,12,1]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).dt.month()
-        )
-
-    def day(self: Self) -> ExprT:
-        """Extract day from underlying DateTime representation.
-
-        Returns the day of month starting from 1. The return value ranges from 1 to 31. (The last day of month differs by months.)
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> from datetime import datetime
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {
-            ...     "datetime": [
-            ...         datetime(1978, 6, 1),
-            ...         datetime(2024, 12, 13),
-            ...         datetime(2065, 1, 1),
-            ...     ]
-            ... }
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_dt_day(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         nw.col("datetime").dt.day().alias("day"),
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_dt_day`:
-
-            >>> agnostic_dt_day(df_pd)
-                datetime  day
-            0 1978-06-01    1
-            1 2024-12-13   13
-            2 2065-01-01    1
-
-            >>> agnostic_dt_day(df_pl)
-            shape: (3, 2)
-            ┌─────────────────────┬─────┐
-            │ datetime            ┆ day │
-            │ ---                 ┆ --- │
-            │ datetime[μs]        ┆ i8  │
-            ╞═════════════════════╪═════╡
-            │ 1978-06-01 00:00:00 ┆ 1   │
-            │ 2024-12-13 00:00:00 ┆ 13  │
-            │ 2065-01-01 00:00:00 ┆ 1   │
-            └─────────────────────┴─────┘
-
-            >>> agnostic_dt_day(df_pa)
-            pyarrow.Table
-            datetime: timestamp[us]
-            day: int64
-            ----
-            datetime: [[1978-06-01 00:00:00.000000,2024-12-13 00:00:00.000000,2065-01-01 00:00:00.000000]]
-            day: [[1,13,1]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).dt.day()
-        )
-
-    def hour(self: Self) -> ExprT:
-        """Extract hour from underlying DateTime representation.
-
-        Returns the hour number from 0 to 23.
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> from datetime import datetime
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {
-            ...     "datetime": [
-            ...         datetime(1978, 1, 1, 1),
-            ...         datetime(2024, 10, 13, 5),
-            ...         datetime(2065, 1, 1, 10),
-            ...     ]
-            ... }
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_dt_hour(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         nw.col("datetime").dt.hour().alias("hour")
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_dt_hour`:
-
-            >>> agnostic_dt_hour(df_pd)
-                         datetime  hour
-            0 1978-01-01 01:00:00     1
-            1 2024-10-13 05:00:00     5
-            2 2065-01-01 10:00:00    10
-
-            >>> agnostic_dt_hour(df_pl)
-            shape: (3, 2)
-            ┌─────────────────────┬──────┐
-            │ datetime            ┆ hour │
-            │ ---                 ┆ ---  │
-            │ datetime[μs]        ┆ i8   │
-            ╞═════════════════════╪══════╡
-            │ 1978-01-01 01:00:00 ┆ 1    │
-            │ 2024-10-13 05:00:00 ┆ 5    │
-            │ 2065-01-01 10:00:00 ┆ 10   │
-            └─────────────────────┴──────┘
-
-            >>> agnostic_dt_hour(df_pa)
-            pyarrow.Table
-            datetime: timestamp[us]
-            hour: int64
-            ----
-            datetime: [[1978-01-01 01:00:00.000000,2024-10-13 05:00:00.000000,2065-01-01 10:00:00.000000]]
-            hour: [[1,5,10]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).dt.hour()
-        )
-
-    def minute(self: Self) -> ExprT:
-        """Extract minutes from underlying DateTime representation.
-
-        Returns the minute number from 0 to 59.
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> from datetime import datetime
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {
-            ...     "datetime": [
-            ...         datetime(1978, 1, 1, 1, 1),
-            ...         datetime(2024, 10, 13, 5, 30),
-            ...         datetime(2065, 1, 1, 10, 20),
-            ...     ]
-            ... }
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_dt_minute(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         nw.col("datetime").dt.minute().alias("minute"),
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_dt_minute`:
-
-            >>> agnostic_dt_minute(df_pd)
-                         datetime  minute
-            0 1978-01-01 01:01:00       1
-            1 2024-10-13 05:30:00      30
-            2 2065-01-01 10:20:00      20
-
-            >>> agnostic_dt_minute(df_pl)
-            shape: (3, 2)
-            ┌─────────────────────┬────────┐
-            │ datetime            ┆ minute │
-            │ ---                 ┆ ---    │
-            │ datetime[μs]        ┆ i8     │
-            ╞═════════════════════╪════════╡
-            │ 1978-01-01 01:01:00 ┆ 1      │
-            │ 2024-10-13 05:30:00 ┆ 30     │
-            │ 2065-01-01 10:20:00 ┆ 20     │
-            └─────────────────────┴────────┘
-
-            >>> agnostic_dt_minute(df_pa)
-            pyarrow.Table
-            datetime: timestamp[us]
-            minute: int64
-            ----
-            datetime: [[1978-01-01 01:01:00.000000,2024-10-13 05:30:00.000000,2065-01-01 10:20:00.000000]]
-            minute: [[1,30,20]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).dt.minute()
-        )
-
-    def second(self: Self) -> ExprT:
-        """Extract seconds from underlying DateTime representation.
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> from datetime import datetime
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {
-            ...     "datetime": [
-            ...         datetime(1978, 1, 1, 1, 1, 1),
-            ...         datetime(2024, 10, 13, 5, 30, 14),
-            ...         datetime(2065, 1, 1, 10, 20, 30),
-            ...     ]
-            ... }
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_dt_second(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         nw.col("datetime").dt.second().alias("second"),
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_dt_second`:
-
-            >>> agnostic_dt_second(df_pd)
-                         datetime  second
-            0 1978-01-01 01:01:01       1
-            1 2024-10-13 05:30:14      14
-            2 2065-01-01 10:20:30      30
-
-            >>> agnostic_dt_second(df_pl)
-            shape: (3, 2)
-            ┌─────────────────────┬────────┐
-            │ datetime            ┆ second │
-            │ ---                 ┆ ---    │
-            │ datetime[μs]        ┆ i8     │
-            ╞═════════════════════╪════════╡
-            │ 1978-01-01 01:01:01 ┆ 1      │
-            │ 2024-10-13 05:30:14 ┆ 14     │
-            │ 2065-01-01 10:20:30 ┆ 30     │
-            └─────────────────────┴────────┘
-
-            >>> agnostic_dt_second(df_pa)
-            pyarrow.Table
-            datetime: timestamp[us]
-            second: int64
-            ----
-            datetime: [[1978-01-01 01:01:01.000000,2024-10-13 05:30:14.000000,2065-01-01 10:20:30.000000]]
-            second: [[1,14,30]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).dt.second()
-        )
-
-    def millisecond(self: Self) -> ExprT:
-        """Extract milliseconds from underlying DateTime representation.
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> from datetime import datetime
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {
-            ...     "datetime": [
-            ...         datetime(1978, 1, 1, 1, 1, 1, 0),
-            ...         datetime(2024, 10, 13, 5, 30, 14, 505000),
-            ...         datetime(2065, 1, 1, 10, 20, 30, 67000),
-            ...     ]
-            ... }
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_dt_millisecond(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         nw.col("datetime").dt.millisecond().alias("millisecond"),
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_dt_millisecond`:
-
-            >>> agnostic_dt_millisecond(df_pd)
-                             datetime  millisecond
-            0 1978-01-01 01:01:01.000            0
-            1 2024-10-13 05:30:14.505          505
-            2 2065-01-01 10:20:30.067           67
-
-            >>> agnostic_dt_millisecond(df_pl)
-            shape: (3, 2)
-            ┌─────────────────────────┬─────────────┐
-            │ datetime                ┆ millisecond │
-            │ ---                     ┆ ---         │
-            │ datetime[μs]            ┆ i32         │
-            ╞═════════════════════════╪═════════════╡
-            │ 1978-01-01 01:01:01     ┆ 0           │
-            │ 2024-10-13 05:30:14.505 ┆ 505         │
-            │ 2065-01-01 10:20:30.067 ┆ 67          │
-            └─────────────────────────┴─────────────┘
-
-            >>> agnostic_dt_millisecond(df_pa)
-            pyarrow.Table
-            datetime: timestamp[us]
-            millisecond: int64
-            ----
-            datetime: [[1978-01-01 01:01:01.000000,2024-10-13 05:30:14.505000,2065-01-01 10:20:30.067000]]
-            millisecond: [[0,505,67]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).dt.millisecond()
-        )
-
-    def microsecond(self: Self) -> ExprT:
-        """Extract microseconds from underlying DateTime representation.
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> from datetime import datetime
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {
-            ...     "datetime": [
-            ...         datetime(1978, 1, 1, 1, 1, 1, 0),
-            ...         datetime(2024, 10, 13, 5, 30, 14, 505000),
-            ...         datetime(2065, 1, 1, 10, 20, 30, 67000),
-            ...     ]
-            ... }
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_dt_microsecond(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         nw.col("datetime").dt.microsecond().alias("microsecond"),
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_dt_microsecond`:
-
-            >>> agnostic_dt_microsecond(df_pd)
-                             datetime  microsecond
-            0 1978-01-01 01:01:01.000            0
-            1 2024-10-13 05:30:14.505       505000
-            2 2065-01-01 10:20:30.067        67000
-
-            >>> agnostic_dt_microsecond(df_pl)
-            shape: (3, 2)
-            ┌─────────────────────────┬─────────────┐
-            │ datetime                ┆ microsecond │
-            │ ---                     ┆ ---         │
-            │ datetime[μs]            ┆ i32         │
-            ╞═════════════════════════╪═════════════╡
-            │ 1978-01-01 01:01:01     ┆ 0           │
-            │ 2024-10-13 05:30:14.505 ┆ 505000      │
-            │ 2065-01-01 10:20:30.067 ┆ 67000       │
-            └─────────────────────────┴─────────────┘
-
-            >>> agnostic_dt_microsecond(df_pa)
-            pyarrow.Table
-            datetime: timestamp[us]
-            microsecond: int64
-            ----
-            datetime: [[1978-01-01 01:01:01.000000,2024-10-13 05:30:14.505000,2065-01-01 10:20:30.067000]]
-            microsecond: [[0,505000,67000]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).dt.microsecond()
-        )
-
-    def nanosecond(self: Self) -> ExprT:
-        """Extract Nanoseconds from underlying DateTime representation.
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> from datetime import datetime
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {
-            ...     "datetime": [
-            ...         datetime(1978, 1, 1, 1, 1, 1, 0),
-            ...         datetime(2024, 10, 13, 5, 30, 14, 500000),
-            ...         datetime(2065, 1, 1, 10, 20, 30, 60000),
-            ...     ]
-            ... }
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_dt_nanosecond(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         nw.col("datetime").dt.nanosecond().alias("nanosecond"),
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_dt_nanosecond`:
-
-            >>> agnostic_dt_nanosecond(df_pd)
-                             datetime  nanosecond
-            0 1978-01-01 01:01:01.000           0
-            1 2024-10-13 05:30:14.500   500000000
-            2 2065-01-01 10:20:30.060    60000000
-
-            >>> agnostic_dt_nanosecond(df_pl)
-            shape: (3, 2)
-            ┌─────────────────────────┬────────────┐
-            │ datetime                ┆ nanosecond │
-            │ ---                     ┆ ---        │
-            │ datetime[μs]            ┆ i32        │
-            ╞═════════════════════════╪════════════╡
-            │ 1978-01-01 01:01:01     ┆ 0          │
-            │ 2024-10-13 05:30:14.500 ┆ 500000000  │
-            │ 2065-01-01 10:20:30.060 ┆ 60000000   │
-            └─────────────────────────┴────────────┘
-
-            >>> agnostic_dt_nanosecond(df_pa)
-            pyarrow.Table
-            datetime: timestamp[us]
-            nanosecond: int64
-            ----
-            datetime: [[1978-01-01 01:01:01.000000,2024-10-13 05:30:14.500000,2065-01-01 10:20:30.060000]]
-            nanosecond: [[0,500000000,60000000]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).dt.nanosecond()
-        )
-
-    def ordinal_day(self: Self) -> ExprT:
-        """Get ordinal day.
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> from datetime import datetime
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {"a": [datetime(2020, 1, 1), datetime(2020, 8, 3)]}
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_dt_ordinal_day(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         a_ordinal_day=nw.col("a").dt.ordinal_day()
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_dt_ordinal_day`:
-
-            >>> agnostic_dt_ordinal_day(df_pd)
-                       a  a_ordinal_day
-            0 2020-01-01              1
-            1 2020-08-03            216
-
-            >>> agnostic_dt_ordinal_day(df_pl)
-            shape: (2, 2)
-            ┌─────────────────────┬───────────────┐
-            │ a                   ┆ a_ordinal_day │
-            │ ---                 ┆ ---           │
-            │ datetime[μs]        ┆ i16           │
-            ╞═════════════════════╪═══════════════╡
-            │ 2020-01-01 00:00:00 ┆ 1             │
-            │ 2020-08-03 00:00:00 ┆ 216           │
-            └─────────────────────┴───────────────┘
-
-            >>> agnostic_dt_ordinal_day(df_pa)
-            pyarrow.Table
-            a: timestamp[us]
-            a_ordinal_day: int64
-            ----
-            a: [[2020-01-01 00:00:00.000000,2020-08-03 00:00:00.000000]]
-            a_ordinal_day: [[1,216]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).dt.ordinal_day()
-        )
-
-    def weekday(self: Self) -> ExprT:
-        """Extract the week day from the underlying Date representation.
-
-        Returns:
-            Returns the ISO weekday number where monday = 1 and sunday = 7
-
-        Examples:
-            >>> from datetime import datetime
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {"a": [datetime(2020, 1, 1), datetime(2020, 8, 3)]}
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_dt_weekday(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(a_weekday=nw.col("a").dt.weekday()).to_native()
-
-            We can then pass either pandas, Polars, PyArrow, and other supported libraries to
-            `agnostic_dt_weekday`:
-
-            >>> agnostic_dt_weekday(df_pd)
-                       a  a_weekday
-            0 2020-01-01          3
-            1 2020-08-03          1
-
-            >>> agnostic_dt_weekday(df_pl)
-            shape: (2, 2)
-            ┌─────────────────────┬───────────┐
-            │ a                   ┆ a_weekday │
-            │ ---                 ┆ ---       │
-            │ datetime[μs]        ┆ i8        │
-            ╞═════════════════════╪═══════════╡
-            │ 2020-01-01 00:00:00 ┆ 3         │
-            │ 2020-08-03 00:00:00 ┆ 1         │
-            └─────────────────────┴───────────┘
-
-            >>> agnostic_dt_weekday(df_pa)
-            pyarrow.Table
-            a: timestamp[us]
-            a_weekday: int64
-            ----
-            a: [[2020-01-01 00:00:00.000000,2020-08-03 00:00:00.000000]]
-            a_weekday: [[3,1]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).dt.weekday()
-        )
-
-    def total_minutes(self: Self) -> ExprT:
-        """Get total minutes.
-
-        Returns:
-            A new expression.
-
-        Notes:
-            The function outputs the total minutes in the int dtype by default,
-            however, pandas may change the dtype to float when there are missing values,
-            consider using `fill_null()` and `cast` in this case.
-
-        Examples:
-            >>> from datetime import timedelta
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {"a": [timedelta(minutes=10), timedelta(minutes=20, seconds=40)]}
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_dt_total_minutes(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         a_total_minutes=nw.col("a").dt.total_minutes()
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_dt_total_minutes`:
-
-            >>> agnostic_dt_total_minutes(df_pd)
-                            a  a_total_minutes
-            0 0 days 00:10:00               10
-            1 0 days 00:20:40               20
-
-            >>> agnostic_dt_total_minutes(df_pl)
-            shape: (2, 2)
-            ┌──────────────┬─────────────────┐
-            │ a            ┆ a_total_minutes │
-            │ ---          ┆ ---             │
-            │ duration[μs] ┆ i64             │
-            ╞══════════════╪═════════════════╡
-            │ 10m          ┆ 10              │
-            │ 20m 40s      ┆ 20              │
-            └──────────────┴─────────────────┘
-
-            >>> agnostic_dt_total_minutes(df_pa)
-            pyarrow.Table
-            a: duration[us]
-            a_total_minutes: int64
-            ----
-            a: [[600000000,1240000000]]
-            a_total_minutes: [[10,20]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).dt.total_minutes()
-        )
-
-    def total_seconds(self: Self) -> ExprT:
-        """Get total seconds.
-
-        Returns:
-            A new expression.
-
-        Notes:
-            The function outputs the total seconds in the int dtype by default,
-            however, pandas may change the dtype to float when there are missing values,
-            consider using `fill_null()` and `cast` in this case.
-
-        Examples:
-            >>> from datetime import timedelta
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {"a": [timedelta(seconds=10), timedelta(seconds=20, milliseconds=40)]}
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_dt_total_seconds(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         a_total_seconds=nw.col("a").dt.total_seconds()
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_dt_total_seconds`:
-
-            >>> agnostic_dt_total_seconds(df_pd)
-                                   a  a_total_seconds
-            0        0 days 00:00:10               10
-            1 0 days 00:00:20.040000               20
-
-            >>> agnostic_dt_total_seconds(df_pl)
-            shape: (2, 2)
-            ┌──────────────┬─────────────────┐
-            │ a            ┆ a_total_seconds │
-            │ ---          ┆ ---             │
-            │ duration[μs] ┆ i64             │
-            ╞══════════════╪═════════════════╡
-            │ 10s          ┆ 10              │
-            │ 20s 40ms     ┆ 20              │
-            └──────────────┴─────────────────┘
-
-            >>> agnostic_dt_total_seconds(df_pa)
-            pyarrow.Table
-            a: duration[us]
-            a_total_seconds: int64
-            ----
-            a: [[10000000,20040000]]
-            a_total_seconds: [[10,20]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).dt.total_seconds()
-        )
-
-    def total_milliseconds(self: Self) -> ExprT:
-        """Get total milliseconds.
-
-        Returns:
-            A new expression.
-
-        Notes:
-            The function outputs the total milliseconds in the int dtype by default,
-            however, pandas may change the dtype to float when there are missing values,
-            consider using `fill_null()` and `cast` in this case.
-
-        Examples:
-            >>> from datetime import timedelta
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {
-            ...     "a": [
-            ...         timedelta(milliseconds=10),
-            ...         timedelta(milliseconds=20, microseconds=40),
-            ...     ]
-            ... }
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_dt_total_milliseconds(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         a_total_milliseconds=nw.col("a").dt.total_milliseconds()
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_dt_total_milliseconds`:
-
-            >>> agnostic_dt_total_milliseconds(df_pd)
-                                   a  a_total_milliseconds
-            0 0 days 00:00:00.010000                    10
-            1 0 days 00:00:00.020040                    20
-
-            >>> agnostic_dt_total_milliseconds(df_pl)
-            shape: (2, 2)
-            ┌──────────────┬──────────────────────┐
-            │ a            ┆ a_total_milliseconds │
-            │ ---          ┆ ---                  │
-            │ duration[μs] ┆ i64                  │
-            ╞══════════════╪══════════════════════╡
-            │ 10ms         ┆ 10                   │
-            │ 20040µs      ┆ 20                   │
-            └──────────────┴──────────────────────┘
-
-            >>> agnostic_dt_total_milliseconds(df_pa)
-            pyarrow.Table
-            a: duration[us]
-            a_total_milliseconds: int64
-            ----
-            a: [[10000,20040]]
-            a_total_milliseconds: [[10,20]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).dt.total_milliseconds()
-        )
-
-    def total_microseconds(self: Self) -> ExprT:
-        """Get total microseconds.
-
-        Returns:
-            A new expression.
-
-        Notes:
-            The function outputs the total microseconds in the int dtype by default,
-            however, pandas may change the dtype to float when there are missing values,
-            consider using `fill_null()` and `cast` in this case.
-
-        Examples:
-            >>> from datetime import timedelta
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {
-            ...     "a": [
-            ...         timedelta(microseconds=10),
-            ...         timedelta(milliseconds=1, microseconds=200),
-            ...     ]
-            ... }
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_dt_total_microseconds(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         a_total_microseconds=nw.col("a").dt.total_microseconds()
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_dt_total_microseconds`:
-
-            >>> agnostic_dt_total_microseconds(df_pd)
-                                   a  a_total_microseconds
-            0 0 days 00:00:00.000010                    10
-            1 0 days 00:00:00.001200                  1200
-
-            >>> agnostic_dt_total_microseconds(df_pl)
-            shape: (2, 2)
-            ┌──────────────┬──────────────────────┐
-            │ a            ┆ a_total_microseconds │
-            │ ---          ┆ ---                  │
-            │ duration[μs] ┆ i64                  │
-            ╞══════════════╪══════════════════════╡
-            │ 10µs         ┆ 10                   │
-            │ 1200µs       ┆ 1200                 │
-            └──────────────┴──────────────────────┘
-
-            >>> agnostic_dt_total_microseconds(df_pa)
-            pyarrow.Table
-            a: duration[us]
-            a_total_microseconds: int64
-            ----
-            a: [[10,1200]]
-            a_total_microseconds: [[10,1200]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).dt.total_microseconds()
-        )
-
-    def total_nanoseconds(self: Self) -> ExprT:
-        """Get total nanoseconds.
-
-        Returns:
-            A new expression.
-
-        Notes:
-            The function outputs the total nanoseconds in the int dtype by default,
-            however, pandas may change the dtype to float when there are missing values,
-            consider using `fill_null()` and `cast` in this case.
-
-        Examples:
-            >>> from datetime import timedelta
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = ["2024-01-01 00:00:00.000000001", "2024-01-01 00:00:00.000000002"]
-            >>> df_pd = pd.DataFrame({"a": pd.to_datetime(data)})
-            >>> df_pl = pl.DataFrame({"a": data}).with_columns(
-            ...     pl.col("a").str.to_datetime(time_unit="ns")
-            ... )
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_dt_total_nanoseconds(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         a_diff_total_nanoseconds=nw.col("a").diff().dt.total_nanoseconds()
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_dt_total_nanoseconds`:
-
-            >>> agnostic_dt_total_nanoseconds(df_pd)
-                                          a  a_diff_total_nanoseconds
-            0 2024-01-01 00:00:00.000000001                       NaN
-            1 2024-01-01 00:00:00.000000002                       1.0
-
-            >>> agnostic_dt_total_nanoseconds(df_pl)
-            shape: (2, 2)
-            ┌───────────────────────────────┬──────────────────────────┐
-            │ a                             ┆ a_diff_total_nanoseconds │
-            │ ---                           ┆ ---                      │
-            │ datetime[ns]                  ┆ i64                      │
-            ╞═══════════════════════════════╪══════════════════════════╡
-            │ 2024-01-01 00:00:00.000000001 ┆ null                     │
-            │ 2024-01-01 00:00:00.000000002 ┆ 1                        │
-            └───────────────────────────────┴──────────────────────────┘
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).dt.total_nanoseconds()
-        )
-
-    def to_string(self: Self, format: str) -> ExprT:  # noqa: A002
-        """Convert a Date/Time/Datetime column into a String column with the given format.
-
-        Arguments:
-            format: Format to format temporal column with.
-
-        Returns:
-            A new expression.
-
-        Notes:
-            Unfortunately, different libraries interpret format directives a bit
-            differently.
-
-            - Chrono, the library used by Polars, uses `"%.f"` for fractional seconds,
-              whereas pandas and Python stdlib use `".%f"`.
-            - PyArrow interprets `"%S"` as "seconds, including fractional seconds"
-              whereas most other tools interpret it as "just seconds, as 2 digits".
-
-            Therefore, we make the following adjustments:
-
-            - for pandas-like libraries, we replace `"%S.%f"` with `"%S%.f"`.
-            - for PyArrow, we replace `"%S.%f"` with `"%S"`.
-
-            Workarounds like these don't make us happy, and we try to avoid them as
-            much as possible, but here we feel like it's the best compromise.
-
-            If you just want to format a date/datetime Series as a local datetime
-            string, and have it work as consistently as possible across libraries,
-            we suggest using:
-
-            - `"%Y-%m-%dT%H:%M:%S%.f"` for datetimes
-            - `"%Y-%m-%d"` for dates
-
-            though note that, even then, different tools may return a different number
-            of trailing zeros. Nonetheless, this is probably consistent enough for
-            most applications.
-
-            If you have an application where this is not enough, please open an issue
-            and let us know.
-
-        Examples:
-            >>> from datetime import datetime
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {
-            ...     "a": [
-            ...         datetime(2020, 3, 1),
-            ...         datetime(2020, 4, 1),
-            ...         datetime(2020, 5, 1),
-            ...     ]
-            ... }
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_dt_to_string(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.select(
-            ...         nw.col("a").dt.to_string("%Y/%m/%d %H:%M:%S")
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_dt_to_string`:
-
-            >>> agnostic_dt_to_string(df_pd)
-                                 a
-            0  2020/03/01 00:00:00
-            1  2020/04/01 00:00:00
-            2  2020/05/01 00:00:00
-
-            >>> agnostic_dt_to_string(df_pl)
-            shape: (3, 1)
-            ┌─────────────────────┐
-            │ a                   │
-            │ ---                 │
-            │ str                 │
-            ╞═════════════════════╡
-            │ 2020/03/01 00:00:00 │
-            │ 2020/04/01 00:00:00 │
-            │ 2020/05/01 00:00:00 │
-            └─────────────────────┘
-
-            >>> agnostic_dt_to_string(df_pa)
-            pyarrow.Table
-            a: string
-            ----
-            a: [["2020/03/01 00:00:00.000000","2020/04/01 00:00:00.000000","2020/05/01 00:00:00.000000"]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).dt.to_string(format)
-        )
-
-    def replace_time_zone(self: Self, time_zone: str | None) -> ExprT:
-        """Replace time zone.
-
-        Arguments:
-            time_zone: Target time zone.
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> from datetime import datetime, timezone
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {
-            ...     "a": [
-            ...         datetime(2024, 1, 1, tzinfo=timezone.utc),
-            ...         datetime(2024, 1, 2, tzinfo=timezone.utc),
-            ...     ]
-            ... }
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            Let's define a dataframe-agnostic function:
-
-            >>> def agnostic_dt_replace_time_zone(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.select(
-            ...         nw.col("a").dt.replace_time_zone("Asia/Kathmandu")
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_dt_replace_time_zone`:
-
-            >>> agnostic_dt_replace_time_zone(df_pd)
-                                      a
-            0 2024-01-01 00:00:00+05:45
-            1 2024-01-02 00:00:00+05:45
-
-            >>> agnostic_dt_replace_time_zone(df_pl)
-            shape: (2, 1)
-            ┌──────────────────────────────┐
-            │ a                            │
-            │ ---                          │
-            │ datetime[μs, Asia/Kathmandu] │
-            ╞══════════════════════════════╡
-            │ 2024-01-01 00:00:00 +0545    │
-            │ 2024-01-02 00:00:00 +0545    │
-            └──────────────────────────────┘
-
-            >>> agnostic_dt_replace_time_zone(df_pa)
-            pyarrow.Table
-            a: timestamp[us, tz=Asia/Kathmandu]
-            ----
-            a: [[2023-12-31 18:15:00.000000Z,2024-01-01 18:15:00.000000Z]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).dt.replace_time_zone(time_zone)
-        )
-
-    def convert_time_zone(self: Self, time_zone: str) -> ExprT:
-        """Convert to a new time zone.
-
-        If converting from a time-zone-naive column, then conversion happens
-        as if converting from UTC.
-
-        Arguments:
-            time_zone: Target time zone.
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> from datetime import datetime, timezone
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {
-            ...     "a": [
-            ...         datetime(2024, 1, 1, tzinfo=timezone.utc),
-            ...         datetime(2024, 1, 2, tzinfo=timezone.utc),
-            ...     ]
-            ... }
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            Let's define a dataframe-agnostic function:
-
-            >>> def agnostic_dt_convert_time_zone(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.select(
-            ...         nw.col("a").dt.convert_time_zone("Asia/Kathmandu")
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_dt_convert_time_zone`:
-
-            >>> agnostic_dt_convert_time_zone(df_pd)
-                                      a
-            0 2024-01-01 05:45:00+05:45
-            1 2024-01-02 05:45:00+05:45
-
-            >>> agnostic_dt_convert_time_zone(df_pl)
-            shape: (2, 1)
-            ┌──────────────────────────────┐
-            │ a                            │
-            │ ---                          │
-            │ datetime[μs, Asia/Kathmandu] │
-            ╞══════════════════════════════╡
-            │ 2024-01-01 05:45:00 +0545    │
-            │ 2024-01-02 05:45:00 +0545    │
-            └──────────────────────────────┘
-
-            >>> agnostic_dt_convert_time_zone(df_pa)
-            pyarrow.Table
-            a: timestamp[us, tz=Asia/Kathmandu]
-            ----
-            a: [[2024-01-01 00:00:00.000000Z,2024-01-02 00:00:00.000000Z]]
-        """
-        if time_zone is None:
-            msg = "Target `time_zone` cannot be `None` in `convert_time_zone`. Please use `replace_time_zone(None)` if you want to remove the time zone."
-            raise TypeError(msg)
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).dt.convert_time_zone(time_zone)
-        )
-
-    def timestamp(self: Self, time_unit: Literal["ns", "us", "ms"] = "us") -> ExprT:
-        """Return a timestamp in the given time unit.
-
-        Arguments:
-            time_unit: {'ns', 'us', 'ms'}
-                Time unit.
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> from datetime import date
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {"date": [date(2001, 1, 1), None, date(2001, 1, 3)]}
-            >>> df_pd = pd.DataFrame(data, dtype="datetime64[ns]")
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            Let's define a dataframe-agnostic function:
-
-            >>> def agnostic_dt_timestamp(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(
-            ...         nw.col("date").dt.timestamp().alias("timestamp_us"),
-            ...         nw.col("date").dt.timestamp("ms").alias("timestamp_ms"),
-            ...     ).to_native()
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_dt_timestamp`:
-
-            >>> agnostic_dt_timestamp(df_pd)
-                    date  timestamp_us  timestamp_ms
-            0 2001-01-01  9.783072e+14  9.783072e+11
-            1        NaT           NaN           NaN
-            2 2001-01-03  9.784800e+14  9.784800e+11
-
-            >>> agnostic_dt_timestamp(df_pl)
-            shape: (3, 3)
-            ┌────────────┬─────────────────┬──────────────┐
-            │ date       ┆ timestamp_us    ┆ timestamp_ms │
-            │ ---        ┆ ---             ┆ ---          │
-            │ date       ┆ i64             ┆ i64          │
-            ╞════════════╪═════════════════╪══════════════╡
-            │ 2001-01-01 ┆ 978307200000000 ┆ 978307200000 │
-            │ null       ┆ null            ┆ null         │
-            │ 2001-01-03 ┆ 978480000000000 ┆ 978480000000 │
-            └────────────┴─────────────────┴──────────────┘
-
-            >>> agnostic_dt_timestamp(df_pa)
-            pyarrow.Table
-            date: date32[day]
-            timestamp_us: int64
-            timestamp_ms: int64
-            ----
-            date: [[2001-01-01,null,2001-01-03]]
-            timestamp_us: [[978307200000000,null,978480000000000]]
-            timestamp_ms: [[978307200000,null,978480000000]]
-        """
-        if time_unit not in {"ns", "us", "ms"}:
-            msg = (
-                "invalid `time_unit`"
-                f"\n\nExpected one of {{'ns', 'us', 'ms'}}, got {time_unit!r}."
-            )
-            raise ValueError(msg)
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).dt.timestamp(time_unit)
-        )
-
-
-class ExprNameNamespace(Generic[ExprT]):
-    def __init__(self: Self, expr: ExprT) -> None:
-        self._expr = expr
-
-    def keep(self: Self) -> ExprT:
-        r"""Keep the original root name of the expression.
-
-        Returns:
-            A new expression.
-
-        Notes:
-            This will undo any previous renaming operations on the expression.
-            Due to implementation constraints, this method can only be called as the last
-            expression in a chain. Only one name operation per expression will work.
-
-        Examples:
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrame
-            >>>
-            >>> data = {"foo": [1, 2], "BAR": [4, 5]}
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_name_keep(df_native: IntoFrame) -> list[str]:
-            ...     df = nw.from_native(df_native)
-            ...     return df.select(nw.col("foo").alias("alias_for_foo").name.keep()).columns
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_name_keep`:
-
-            >>> agnostic_name_keep(df_pd)
-            ['foo']
-
-            >>> agnostic_name_keep(df_pl)
-            ['foo']
-
-            >>> agnostic_name_keep(df_pa)
-            ['foo']
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).name.keep()
-        )
-
-    def map(self: Self, function: Callable[[str], str]) -> ExprT:
-        r"""Rename the output of an expression by mapping a function over the root name.
-
-        Arguments:
-            function: Function that maps a root name to a new name.
-
-        Returns:
-            A new expression.
-
-        Notes:
-            This will undo any previous renaming operations on the expression.
-            Due to implementation constraints, this method can only be called as the last
-            expression in a chain. Only one name operation per expression will work.
-
-        Examples:
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrame
-            >>>
-            >>> data = {"foo": [1, 2], "BAR": [4, 5]}
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> renaming_func = lambda s: s[::-1]  # reverse column name
-            >>> def agnostic_name_map(df_native: IntoFrame) -> list[str]:
-            ...     df = nw.from_native(df_native)
-            ...     return df.select(nw.col("foo", "BAR").name.map(renaming_func)).columns
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_name_map`:
-
-            >>> agnostic_name_map(df_pd)
-            ['oof', 'RAB']
-
-            >>> agnostic_name_map(df_pl)
-            ['oof', 'RAB']
-
-            >>> agnostic_name_map(df_pa)
-            ['oof', 'RAB']
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).name.map(function)
-        )
-
-    def prefix(self: Self, prefix: str) -> ExprT:
-        r"""Add a prefix to the root column name of the expression.
-
-        Arguments:
-            prefix: Prefix to add to the root column name.
-
-        Returns:
-            A new expression.
-
-        Notes:
-            This will undo any previous renaming operations on the expression.
-            Due to implementation constraints, this method can only be called as the last
-            expression in a chain. Only one name operation per expression will work.
-
-        Examples:
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrame
-            >>>
-            >>> data = {"foo": [1, 2], "BAR": [4, 5]}
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_name_prefix(df_native: IntoFrame, prefix: str) -> list[str]:
-            ...     df = nw.from_native(df_native)
-            ...     return df.select(nw.col("foo", "BAR").name.prefix(prefix)).columns
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_name_prefix`:
-
-            >>> agnostic_name_prefix(df_pd, "with_prefix_")
-            ['with_prefix_foo', 'with_prefix_BAR']
-
-            >>> agnostic_name_prefix(df_pl, "with_prefix_")
-            ['with_prefix_foo', 'with_prefix_BAR']
-
-            >>> agnostic_name_prefix(df_pa, "with_prefix_")
-            ['with_prefix_foo', 'with_prefix_BAR']
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).name.prefix(prefix)
-        )
-
-    def suffix(self: Self, suffix: str) -> ExprT:
-        r"""Add a suffix to the root column name of the expression.
-
-        Arguments:
-            suffix: Suffix to add to the root column name.
-
-        Returns:
-            A new expression.
-
-        Notes:
-            This will undo any previous renaming operations on the expression.
-            Due to implementation constraints, this method can only be called as the last
-            expression in a chain. Only one name operation per expression will work.
-
-        Examples:
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrame
-            >>>
-            >>> data = {"foo": [1, 2], "BAR": [4, 5]}
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_name_suffix(df_native: IntoFrame, suffix: str) -> list[str]:
-            ...     df = nw.from_native(df_native)
-            ...     return df.select(nw.col("foo", "BAR").name.suffix(suffix)).columns
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_name_suffix`:
-
-            >>> agnostic_name_suffix(df_pd, "_with_suffix")
-            ['foo_with_suffix', 'BAR_with_suffix']
-
-            >>> agnostic_name_suffix(df_pl, "_with_suffix")
-            ['foo_with_suffix', 'BAR_with_suffix']
-
-            >>> agnostic_name_suffix(df_pa, "_with_suffix")
-            ['foo_with_suffix', 'BAR_with_suffix']
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).name.suffix(suffix)
-        )
-
-    def to_lowercase(self: Self) -> ExprT:
-        r"""Make the root column name lowercase.
-
-        Returns:
-            A new expression.
-
-        Notes:
-            This will undo any previous renaming operations on the expression.
-            Due to implementation constraints, this method can only be called as the last
-            expression in a chain. Only one name operation per expression will work.
-
-        Examples:
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrame
-            >>>
-            >>> data = {"foo": [1, 2], "BAR": [4, 5]}
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_name_to_lowercase(df_native: IntoFrame) -> list[str]:
-            ...     df = nw.from_native(df_native)
-            ...     return df.select(nw.col("foo", "BAR").name.to_lowercase()).columns
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_name_to_lowercase`:
-
-            >>> agnostic_name_to_lowercase(df_pd)
-            ['foo', 'bar']
-
-            >>> agnostic_name_to_lowercase(df_pl)
-            ['foo', 'bar']
-
-            >>> agnostic_name_to_lowercase(df_pa)
-            ['foo', 'bar']
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).name.to_lowercase()
-        )
-
-    def to_uppercase(self: Self) -> ExprT:
-        r"""Make the root column name uppercase.
-
-        Returns:
-            A new expression.
-
-        Notes:
-            This will undo any previous renaming operations on the expression.
-            Due to implementation constraints, this method can only be called as the last
-            expression in a chain. Only one name operation per expression will work.
-
-        Examples:
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrame
-            >>>
-            >>> data = {"foo": [1, 2], "BAR": [4, 5]}
-            >>> df_pd = pd.DataFrame(data)
-            >>> df_pl = pl.DataFrame(data)
-            >>> df_pa = pa.table(data)
-
-            We define a dataframe-agnostic function:
-
-            >>> def agnostic_name_to_uppercase(df_native: IntoFrame) -> list[str]:
-            ...     df = nw.from_native(df_native)
-            ...     return df.select(nw.col("foo", "BAR").name.to_uppercase()).columns
-
-            We can then pass any supported library such as pandas, Polars, or
-            PyArrow to `agnostic_name_to_uppercase`:
-
-            >>> agnostic_name_to_uppercase(df_pd)
-            ['FOO', 'BAR']
-
-            >>> agnostic_name_to_uppercase(df_pl)
-            ['FOO', 'BAR']
-
-            >>> agnostic_name_to_uppercase(df_pa)
-            ['FOO', 'BAR']
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).name.to_uppercase()
-        )
-
-
-class ExprListNamespace(Generic[ExprT]):
-    def __init__(self: Self, expr: ExprT) -> None:
-        self._expr = expr
-
-    def len(self: Self) -> ExprT:
-        """Return the number of elements in each list.
-
-        Null values count towards the total.
-
-        Returns:
-            A new expression.
-
-        Examples:
-            >>> import pandas as pd
-            >>> import polars as pl
-            >>> import pyarrow as pa
-            >>> import narwhals as nw
-            >>> from narwhals.typing import IntoFrameT
-            >>>
-            >>> data = {"a": [[1, 2], [3, 4, None], None, []]}
-
-            Let's define a dataframe-agnostic function:
-
-            >>> def agnostic_list_len(df_native: IntoFrameT) -> IntoFrameT:
-            ...     df = nw.from_native(df_native)
-            ...     return df.with_columns(a_len=nw.col("a").list.len()).to_native()
-
-            We can then pass pandas / PyArrow / Polars / any other supported library:
-
-            >>> agnostic_list_len(
-            ...     pd.DataFrame(data).astype({"a": pd.ArrowDtype(pa.list_(pa.int64()))})
-            ... )  # doctest: +SKIP
-                           a  a_len
-            0        [1. 2.]      2
-            1  [ 3.  4. nan]      3
-            2           <NA>   <NA>
-            3             []      0
-
-            >>> agnostic_list_len(pl.DataFrame(data))
-            shape: (4, 2)
-            ┌──────────────┬───────┐
-            │ a            ┆ a_len │
-            │ ---          ┆ ---   │
-            │ list[i64]    ┆ u32   │
-            ╞══════════════╪═══════╡
-            │ [1, 2]       ┆ 2     │
-            │ [3, 4, null] ┆ 3     │
-            │ null         ┆ null  │
-            │ []           ┆ 0     │
-            └──────────────┴───────┘
-
-            >>> agnostic_list_len(pa.table(data))
-            pyarrow.Table
-            a: list<item: int64>
-              child 0, item: int64
-            a_len: uint32
-            ----
-            a: [[[1,2],[3,4,null],null,[]]]
-            a_len: [[2,3,null,0]]
-        """
-        return self._expr.__class__(
-            lambda plx: self._expr._to_compliant_expr(plx).list.len()
-        )
 
 
 def col(*names: str | Iterable[str]) -> Expr:
